+++ hal_spi.h
@@ -2,4 +2,4 @@
 
 void SPI_Transmit(unsigned char data);
 unsigned char SPI_Receive();
-void SPI_Init(uint8_t bits, uint32_t speed, uint16_t delay);
+void SPI_Init(int bits, int speed, int delay);
